Move rest tests (#36052)
