+++ src/rest/tests/rendering.js
@@ -1,49 +1,13 @@
 import { jest, test } from '@jest/globals'
 import { slug } from 'github-slugger'
-import { readdirSync, readFileSync } from 'fs'
-import path from 'path'
 
-import { get, getDOM } from '../helpers/e2etest.js'
-import getRest, { REST_DATA_DIR, REST_SCHEMA_FILENAME } from '../../src/rest/lib/index.js'
-import { isApiVersioned, allVersions } from '../../lib/all-versions.js'
-import { getDiffOpenAPIContentRest } from '../../src/rest/scripts/test-open-api-schema.js'
+import { getDOM } from '../../../tests/helpers/e2etest.js'
+import getRest from '../lib/index.js'
+import { isApiVersioned, allVersions } from '../../../lib/all-versions.js'
+import { getDiffOpenAPIContentRest } from '../scripts/test-open-api-schema.js'
 
 describe('REST references docs', () => {
   jest.setTimeout(3 * 60 * 1000)
-
-  test('all category and subcategory REST pages render for free-pro-team', async () => {
-    // This currently just grabs the 'free-pro-team' schema, but ideally, we'd
-    // get a list of all categories across all versions.
-    const freeProTeamVersion = readdirSync(REST_DATA_DIR)
-      .filter((file) => file.startsWith('fpt'))
-      .shift()
-    const freeProTeamSchema = JSON.parse(
-      readFileSync(path.join(REST_DATA_DIR, freeProTeamVersion, REST_SCHEMA_FILENAME), 'utf8')
-    )
-
-    const restCategories = Object.entries(freeProTeamSchema)
-      .map(([key, subCategory]) => {
-        const subCategoryKeys = Object.keys(subCategory)
-        if (subCategoryKeys.length === 1) {
-          return key
-        } else {
-          return subCategoryKeys.map((elem) => `${key}/${elem}`)
-        }
-      })
-      .flat()
-
-    const statusCodes = await Promise.all(
-      restCategories.map(async (page) => {
-        const url = `/en/rest/${page}`
-        const res = await get(url)
-        return [url, res.statusCode]
-      })
-    )
-    for (const [url, status] of statusCodes) {
-      expect(status, url).toBe(200)
-    }
-    expect.assertions(restCategories.length)
-  })
 
   // Checks that every version of the /rest/checks
   // page has every operation defined in the openapi schema.
